+++ examples/lesson3_collision.py
@@ -1,6 +1,6 @@
 from kaypy import *
 
-kaplay(width=800, height=600, background=[0, 0, 0])
+kaypy(width=800, height=600, background=[0, 0, 0])
 
 loadSprite("bean", "images/bean.png")
 loadSprite("ghosty", "images/ghosty.png")