+++ examples/lesson6_sprite_animation.py
@@ -1,6 +1,6 @@
 from kaypy import *
 
-kaplay(width=800, height=600, background=[0, 0, 0])
+kaypy(width=800, height=600, background=[0, 0, 0])
 
 loadSprite("dino", [
     "images/dino_0.png", "images/dino_1.png", "images/dino_2.png",